+++ app/src/routes/(auth)/search/+page.svelte
