+++ app/src/routes/login/+page.server.ts
@@ -3,5 +3,5 @@ import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ parent }) => {
   const { user_info } = await parent();
-  if (user_info) redirect(307, '/user');
+  if (user_info) redirect(307, '/');
 };
