--- app/src/routes/parivartak/+page.server.ts
@@ -1,5 +1,0 @@
-import { redirect } from '@sveltejs/kit';
-
-export const load = () => {
-  throw redirect(302, 'https://lipilekhika.in/app');
-};
